replace batch_transform.py
