+++ README.md
@@ -456,7 +456,7 @@ optimizer = torch.optim.Adam(model.parameters())
 scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, [2])
 
 # model training
-runner = dl.SupervisedRunner()
+runner = dl.SupervisedRunner(output_key="logits")
 runner.train(
     model=model,
     criterion=criterion,
@@ -466,7 +466,12 @@ runner.train(
     num_epochs=3,
     verbose=True,
     callbacks=[
-        dl.LambdaPreprocessCallback(input_key="logits", output_key="scores", lambda_fn=torch.sigmoid),
+        dl.BatchTransformCallback(
+            lambda_fn=torch.sigmoid,
+            scope="on_batch_end",
+            input_key="logits",
+            output_key="scores"
+        ),
         dl.CriterionCallback(input_key="logits", target_key="targets", metric_key="loss"),
         # uncomment for extra metrics:
 #         dl.AUCCallback(input_key="scores", target_key="targets"),
