+++ catalyst/callbacks/tests/test_transform_kornia.py
@@ -81,7 +81,13 @@ def test_transform_kornia():
         verbose=False,
         load_best_on_end=True,
         check=True,
-        callbacks=[BatchTransformCallback(transrorms, input_key=0)],
+        callbacks=[
+            BatchTransformCallback(
+                lambda_fn=transrorms,
+                scope="on_batch_start",
+                input_key=0
+            )
+        ],
     )
 
     # model inference
