+++ docs/api/callbacks.rst
@@ -50,13 +50,6 @@ CriterionCallback
     :exclude-members: __init__, on_experiment_start, on_stage_start, on_epoch_start, on_loader_start, on_batch_start, on_batch_end, on_loader_end, on_epoch_end, on_stage_end, on_experiment_end
     :show-inheritance:
 
-LambdaWrapperCallback
-~~~~~~~~~~~~~~~~~~~~~
-.. autoclass:: catalyst.callbacks.lambda_preprocess.LambdaPreprocessCallback
-    :members:
-    :exclude-members: __init__, on_experiment_start, on_stage_start, on_epoch_start, on_loader_start, on_batch_start, on_batch_end, on_loader_end, on_epoch_end, on_stage_end, on_experiment_end
-    :show-inheritance:
-
 Metric – BatchMetricCallback
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. autoclass:: catalyst.callbacks.metric.BatchMetricCallback
